catalog/requirements.txt
aiopg==0.16.0
peewee-async==0.6.0a
yoyo-migrations==6.1.0
fastapi==0.44.1
uvicorn==0.9.0
pytest-asyncio==0.10.0
requests==2.7.0

